+++ packages/core/android/src/main/java/io/sentry/react/RNSentryStart.java
@@ -202,6 +202,13 @@ final class RNSentryStart {
     if (rnOptions.hasKey("enableAnrFingerprinting")) {
       options.setEnableAnrFingerprinting(rnOptions.getBoolean("enableAnrFingerprinting"));
     }
+    if (rnOptions.hasKey("enableNdkAppHangTracking")) {
+      options.setEnableNdkAppHangTracking(rnOptions.getBoolean("enableNdkAppHangTracking"));
+    }
+    if (rnOptions.hasKey("ndkAppHangTimeoutIntervalMillis")) {
+      options.setNdkAppHangTimeoutIntervalMillis(
+          rnOptions.getInt("ndkAppHangTimeoutIntervalMillis"));
+    }
     if (rnOptions.hasKey("spotlight")) {
       if (rnOptions.getType("spotlight") == ReadableType.Boolean) {
         options.setEnableSpotlight(rnOptions.getBoolean("spotlight"));